# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pepsi PM API
  version: 0.8.0
  description: |
    REST API — backend Node/Express ที่ `backend/`
    พอร์ตล็อก dev: **5000** (สอดภาคผนวก ค / INFRASTRUCTURE)
    Auth: `SKIP_AUTH=true` ใน dev ไม่ต้องส่ง Bearer; production ใช้ JWT HS256 + โหลดสิทธิ์จาก `pepsi_pm`
servers:
  - url: http://127.0.0.1:5000
    description: Local backend
...
paths:
...
  /api/v1/work-orders:
    get:
      tags: [WorkOrders]
      summary: List work orders (paged)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paged list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkOrder'
                  total:
                    type: integer
  /api/v1/work-orders/{workOrderId}:
    get:
      tags: [WorkOrders]
      summary: Get one work order by id
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/WorkOrderDetail'
                  requestId:
                    type: string
        '400':
          description: Invalid workOrderId
        '401':
          description: Unauthorized
        '404':
          description: Not found
  /api/v1/work-orders/{workOrderId}/task-logs:
    post:
      tags: [TaskLogs]
      summary: สร้าง task log (ก่อนอัปโหลดรูปหลักฐาน)
      description: |
        สร้างแถว `task_logs` สำหรับใบงานที่มีอยู่ — จากนั้นเรียก `POST /task-logs/{id}/attachments`.
        เมื่อ `SKIP_AUTH=true` และ user stub id=0 จะใช้ `users.gpid = __system__` เป็น `created_by_user_id`.
        ต้องมีสิทธิ์ **`work_order.edit`** เมื่อ `SKIP_AUTH=false`.
      parameters:
        - name: workOrderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                logType:
                  type: string
                  default: photo
                  maxLength: 32
      responses:
        '201':
          description: สร้าง task_logs แล้ว
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskLogCreated'
        '400':
          description: workOrderId ไม่ถูกต้อง / validation
        '401':
          description: Unauthorized
        '403':
          description: ไม่มีสิทธิ์ work_order.edit
        '404':
          description: ไม่พบ work_orders
...
components:
  schemas:
...
    WorkOrder:
      type: object
      properties:
        id:
          type: string
        order_number:
          type: string
        order_type:
          type: string
        system_status:
          type: string
    WorkOrderDetail:
      type: object
      description: แถวเดียวจาก work_orders (รายละเอียดสำหรับหน้า FE)
      properties:
        id:
          type: string
        order_number:
          type: string
        order_type:
          type: string
          nullable: true
        system_status:
          type: string
          nullable: true
        user_status:
          type: string
          nullable: true
        equipment_id:
          type: string
          nullable: true
        work_center_planned:
          type: string
          nullable: true
        work_center_actual:
          type: string
          nullable: true
        planned_start:
          type: string
          format: date-time
          nullable: true
        planned_finish:
          type: string
          format: date-time
          nullable: true
        basic_start:
          type: string
          format: date-time
          nullable: true
        basic_finish:
          type: string
          format: date-time
          nullable: true
        last_import_batch_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
...
    TaskLogCreated:
      type: object
      properties:
        id:
          type: string
          description: PK `task_logs.id`
        workOrderId:
          type: string
        logType:
          type: string
        requestId:
          type: string
...